acara done

+++ app/Http/Controllers/TagnameController.php
@@ -13,16 +13,19 @@ class TagnameController extends Controller
     
     public function read()
     {
+        if(Auth::user()->role != '1') return redirect('/');
         return view('member.tag.read', [
             'tags' => tagname::all(),
          ]);
     }
     public function createView()
     {
+        if(Auth::user()->role != '1') return redirect('/');
         return view('member.tag.create');
     }
     public function create(Request $request)
     {
+        if(Auth::user()->role != '1') return redirect('/');
         $validated = $request->validate([
             'name' => 'required|unique:tagnames|max:255',
             'slug' => 'required|unique:tagnames|max:255',
@@ -32,10 +35,12 @@ class TagnameController extends Controller
         // return redirect($request->fromPage == route('member.berita.tag') ? route('member.berita.tag') : route('member.berita.tag'));
     }
     public function checkSlug(Request $request){
+        if(Auth::user()->role != '1') return redirect('/');
         $slug = SlugService::createSlug(tagname::class, 'slug', $request->title);
         return response()->json(['slug' => $slug]);
     }
     public function newTag(Request $request){
+        if(Auth::user()->role != '1') return redirect('/');
         $request['slug'] = SlugService::createSlug(tagname::class, 'slug', $request->name);
         $result = $request->validate([
             'name' => 'required|unique:categories|max:255',
@@ -46,6 +51,7 @@ class TagnameController extends Controller
     }
     
     public function addTag(Request $request){
+        if(Auth::user()->role != '1') return redirect('/');
         $result = $request->validate([
             'post_id' => 'required',
             'tagname_id'=> 'required'
@@ -58,12 +64,14 @@ class TagnameController extends Controller
         ]);
     }
     public function getTag(Request $request){
+        if(Auth::user()->role != '1') return redirect('/');
         $tagname = tag::with(['tagname', 'post'])->where('post_id', $request->post_id)->get();
         return response()->json([
             'tags' => $tagname, 
         ]);
     }
     public function deleteTag(Request $request){
+        if(Auth::user()->role != '1') return redirect('/');
         $tag = tag::with(['tagname', 'post'])->where('id', $request->id)->first();
         if($tag->post->user_id == Auth::user()->id){
             $tag->delete();
